modified:   workspaces/DummyRosterWebApi/DummyRoster.Common.EntityModel/Models/Invoice.cs

+++ workspaces/DummyRosterWebApi/DummyRoster.Common.EntityModel/Models/Invoice.cs
@@ -15,17 +15,17 @@ public partial class Invoice
 
     [Key]
     [Column(TypeName = "INT")]
-    public long FormId { get; set; }
+    public int FormId { get; set; }
 
     [Key]
     [Column(TypeName = "INT")]
-    public long ProductId { get; set; }
+    public int ProductId { get; set; }
 
     [Column(TypeName = "NUMERIC")]
     public byte[] UnitPrice { get; set; } = null!;
 
     [Column(TypeName = "SMALLINT")]
-    public long Quantity { get; set; }
+    public int Quantity { get; set; }
 
     [Column(TypeName = "NUMERIC")]
     public byte[] PriceCut { get; set; } = null!;
